phase2_Intermediate_Projects/requirements.txt
streamlit
requests
feedparser
openai-whisper
Pillow
transformers
torch
Pillow

